[rename]: CategoryDropdown 컴포넌트 이동 (#82)
- 기존: components/feature/
- 변경: layouts/App/Header/CategoryButton/
